move deepParseJson to utils
